+++ R/lms.R
@@ -1,10 +1,10 @@
-#' Summary for data.frame with replicated rows
-#' @description transforming summary.lm object according to real number of d.f.
+#' Adjusted summary for linear regression when based on replicated dataset
+#' @description adjusting lm object results according to original number of degree of freedom.
 #' The standard errors, t statistics and p values have to be adjusted because of replicated rows.
 #' @param x lm object
 #' @param df_old integer number of d.f in original dataset. For bigger datasets `nrow` should be sufficient.
 #' @param df_new integer number of d.f in dataset with replicated rows, Default: x$df.residual
-#' @return data.frame with additional columns over a regular summary.lm output like correct and statistics adjusted by it.
+#' @return data.frame with additional columns over a regular summary.lm output, like correct and statistics adjusted by it.
 #' @importFrom stats pt
 #' @details The size of the correction is equal to sqrt(df_new / df_old).
 #' Where standard errors are multiplied and t statistics divided by it.