+++ R/plot_cat2cat.R
@@ -1,8 +1,9 @@
-#' Summary plots for cat2cat results.
-#' @description This function help to understand properties of cat2cat resuls and possibly before further preprocessing.
-#' @param data data.frame
-#' @param weis character - name of a certain wei_*_c2c column added by cat2cat function. Default wei_freq_c2c
-#' @param type character - one of 3 types both, hist, bar
+#' Summary plots for cat2cat results
+#' @description This function help to understand properties of cat2cat results.
+#' It is recommended to run it before further processing, like next iterations.
+#' @param data data.frame - one of the data.frames returned by the `cat2cat` function.
+#' @param weis character - name of a certain wei_*_c2c column, added by cat2cat function. Default wei_freq_c2c
+#' @param type character - one of 3 types `"both"`, `"hist"`, `"bar"`.
 #' @return base plot graphics
 #' @importFrom graphics hist barplot par
 #' @importFrom utils head
@@ -21,11 +22,11 @@
 #' plot_c2c(occup_2$old, type = c("both"))
 #' plot_c2c(occup_2$old, type = c("hist"))
 #' plot_c2c(occup_2$old, type = c("bar"))
-plot_c2c <- function(data, weis = "wei_freq_c2c", type = "both") {
+plot_c2c <- function(data, weis = "wei_freq_c2c", type = c("both", "hist", "bar")) {
   assert_that(is.data.frame(data))
   assert_that(!is.null(data[[weis]]), msg = paste("There are no cat2cat additional columns/variables like", weis))
   assert_that(nrow(data) != sum(data[[weis]]), msg = "There are no replications. Probably you should switch between old/new period.")
-  assert_that(length(type) == 1 && (type %in% c("both", "hist", "bar")))
+  type <- match.arg(type)
 
   if (type == "both") {
     default_mfrow <- par("mfrow")