+++ NEWS.md
@@ -1,10 +1,14 @@
-# cat2cat 0.2.2
+# cat2cat 0.3.0
 
 * fixed the `cat2cat` function, the ml part is assuming that categorical variable is always named "code". 
 * fixed problems when using only one feature in the ml part of `cat2cat` function.
-* transfer caret and `randomForest` packages to Suggests, they are delayed loaded now.
+* persist NA values in a categorical variable in the cat2cat function.
+* update vignette.
+* improved roxygen2 docs.
+* transferred caret and `randomForest` packages to Suggests, they are delayed loaded now.
 * styler of the code.
-* remove data.table from Imports.
+* more readable code.
+* removed data.table from Imports.
 
 # cat2cat 0.2.1